Converted to int to reduce typecasting

+++ src/com/frc4343/robot2/Systems/JoystickSystem.java
@@ -16,7 +16,7 @@ public class JoystickSystem extends System {
         }
     }
 
-    public Joystick getJoystick(byte joystickNumber) {
+    public Joystick getJoystick(int joystickNumber) {
         return joysticks[joystickNumber - 1];
     }
 }
